Add files via upload
In our project, we aimed to investigate the complex conditions surrounding heart disease, a
leading cause of morbidity and mortality worldwide. With the increasing availability of
electronic health records and advancements in machine learning algorithms, we sought to
develop a predictive model that can accurately forecast the likelihood of heart disease
based on a set of attributes. By leveraging a comprehensive dataset, our goal was to identify
the most relevant factors contributing to heart disease and create a reliable prediction tool
that can assist healthcare professionals in making informed decisions. This report outlines
our methodology, results, and insights gained from this project.